+++ crates/module-system/module-implementations/integration-tests/src/sequencer_rewards/mod.rs
@@ -9,12 +9,12 @@ use sov_modules_stf_blueprint::{BatchSequencerOutcome, TxEffect};
 use sov_test_utils::auth::TestAuth;
 use sov_test_utils::generators::value_setter::ValueSetterMessages;
 use sov_test_utils::runtime::TestRuntime;
-use sov_test_utils::{new_test_blob_from_batch, MessageGenerator};
-
-use crate::helpers::{
-    AttesterIncentivesParams, BankParams, SequencerParams, TestRollup, DEFAULT_STAKE_AMOUNT,
-    DEFAULT_USER_BALANCE, S,
+use sov_test_utils::{
+    new_test_blob_from_batch, MessageGenerator, TEST_DEFAULT_GAS_LIMIT, TEST_DEFAULT_USER_BALANCE,
+    TEST_DEFAULT_USER_STAKE,
 };
+
+use crate::helpers::{AttesterIncentivesParams, BankParams, SequencerParams, TestRollup, S};
 
 const TEST_PRIORITY_FEE: PriorityFeeBips = PriorityFeeBips::from_percentage(10);
 const NUM_TXS_PER_BATCH: u64 = 2;
@@ -66,16 +66,16 @@ fn test_sequencer_reward_in_stf(rollup: &mut TestRollup, max_fee: u64) -> Result
     let seq_rollup_addr = seq_params.rollup_address;
     let seq_da_addr = seq_params.da_address;
     let bank_params = BankParams::with_addresses_and_balances(vec![
-        (seq_params.rollup_address, DEFAULT_USER_BALANCE),
-        (admin_pub_key, DEFAULT_USER_BALANCE),
+        (seq_params.rollup_address, TEST_DEFAULT_USER_BALANCE),
+        (admin_pub_key, TEST_DEFAULT_USER_BALANCE),
     ]);
     let attester_params = AttesterIncentivesParams::default();
 
     // Genesis
     let init_root_hash = rollup.genesis(admin_pub_key, seq_params, bank_params, attester_params);
 
-    let post_genesis_sequencer_balance = DEFAULT_USER_BALANCE - DEFAULT_STAKE_AMOUNT;
-    let post_genesis_registry_balance = DEFAULT_STAKE_AMOUNT;
+    let post_genesis_sequencer_balance = TEST_DEFAULT_USER_BALANCE - TEST_DEFAULT_USER_STAKE;
+    let post_genesis_registry_balance = TEST_DEFAULT_USER_STAKE;
 
     check_sequencer_and_registry_balances(
         rollup,
@@ -150,7 +150,7 @@ fn test_sequencer_rewarded_max_priority_fee() -> Result<(), Infallible> {
     let mut rollup = TestRollup::new();
 
     // The max fee is the same as the base fee so the sequencer should not get rewarded
-    let max_fee = <S as Spec>::Gas::from_slice(&[1_000_000, 1_000_000])
+    let max_fee = <S as Spec>::Gas::from_slice(&TEST_DEFAULT_GAS_LIMIT)
         .value(&rollup.initial_base_fee_per_gas());
 
     test_sequencer_reward_in_stf(&mut rollup, max_fee)